修改方法参数，数据类型
- (void)startClockAtHour:(NSUInteger)hour minute:(NSUInteger)minute
second:(NSUInteger)second

+++ Demo/Clock/SLClockView/SLClockView.h
@@ -17,7 +17,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (void)startClockAtNow;
 
-- (void)startClockAtHour:(CGFloat)hour minute:(CGFloat)minute second:(CGFloat)second;
+- (void)startClockAtHour:(NSUInteger)hour minute:(NSUInteger)minute second:(NSUInteger)second;
 
 - (void)startClockAtDate:(NSDate *)date;
